// x-pack/platform/plugins/shared/observability_ai_assistant/public/types.d.ts
export interface ObservabilityAIAssistantChatService {
    sendAnalyticsEvent: (event: TelemetryEventTypeWithPayload) => void;
    chat: (name: string, options: {
        systemMessage: string;
        messages: Message[];
        connectorId: string;
        functions?: Array<Pick<FunctionDefinition, 'name' | 'description' | 'parameters'>>;
        functionCall?: string;
        signal: AbortSignal;
        scopes: AssistantScope[];
    }) => Observable<ChatCompletionChunkEvent>;
    complete: (options: {
        getScreenContexts: () => ObservabilityAIAssistantScreenContext[];
        conversationId?: string;
        connectorId: string;
        systemMessage?: string;
        messages: Message[];
        persist: boolean;
        disableFunctions: boolean;
        signal: AbortSignal;
        instructions?: Array<string | Instruction>;
        scopes: AssistantScope[];
    }) => Observable<StreamingChatResponseEventWithoutError>;
    getFunctions: (options?: {
        contexts?: string[];
        filter?: string;
        scopes: AssistantScope[];
    }) => FunctionDefinition[];
    functions$: BehaviorSubject<FunctionDefinition[]>;
    hasFunction: (name: string) => boolean;
    getSystemMessage: () => string;
    hasRenderFunction: (name: string) => boolean;
    renderFunction: (name: string, args: string | undefined, response: {
        data?: string;
        content?: string;
    }, onActionClick: ChatActionClickHandler) => React.ReactNode;
    getScopes: () => AssistantScope[];
}
export interface ObservabilityAIAssistantConversationService {
    openNewConversation: ({}: {
        messages: Message[];
        title?: string;
        hideConversationList?: boolean;
    }) => void;
    predefinedConversation$: Observable<{
        messages: Message[];
        title?: string;
        hideConversationList?: boolean;
    }>;
}
export interface ObservabilityAIAssistantService {
    callApi: ObservabilityAIAssistantAPIClient;
    isEnabled: () => boolean;
    start: ({}: {
        signal: AbortSignal;
    }) => Promise<ObservabilityAIAssistantChatService>;
    register: (fn: ChatRegistrationRenderFunction) => void;
    setScreenContext: (screenContext: ObservabilityAIAssistantScreenContext) => () => void;
    getScreenContexts: () => ObservabilityAIAssistantScreenContext[];
    conversations: ObservabilityAIAssistantConversationService;
    navigate: (callback: () => void) => Promise<Observable<MessageAddEvent>>;
    scope$: BehaviorSubject<AssistantScope[]>;
    setScopes: (scope: AssistantScope[]) => void;
    getScopes: () => AssistantScope[];
}
export type RenderFunction<TArguments, TResponse extends FunctionResponse> = (options: {
    arguments: TArguments;
    response: TResponse;
    onActionClick: ChatActionClickHandler;
}) => React.ReactNode;
export type RegisterRenderFunctionDefinition<TFunctionArguments = any, TFunctionResponse extends FunctionResponse = any> = (name: string, render: RenderFunction<TFunctionArguments, TFunctionResponse>) => void;
export type ChatRegistrationRenderFunction = ({}: {
    registerRenderFunction: RegisterRenderFunctionDefinition;
}) => Promise<void>;